+++ src/components/pages/media/songs.jsx
@@ -1,0 +1,19 @@
+const songs = [
+  {
+    song: 'Guitar',
+    artist: 'Mo Mo',
+    image: 'https://images.unsplash.com/photo-1513277179122-48cc37a9cd96?w=800&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8OTB8fGNhdHN8ZW58MHx8MHx8fDA%3D',
+  },
+  {
+    song: 'In Slow Motion Inspiring Ambient Lounge',
+    artist: 'One Identifier',
+    image: 'https://images.unsplash.com/photo-1513277179122-48cc37a9cd96?w=800&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8OTB8fGNhdHN8ZW58MHx8MHx8fDA%3D',
+  },
+  {
+    song: 'To Meet the Sun',
+    artist: '42 Nights',
+    image: 'https://images.unsplash.com/photo-1513277179122-48cc37a9cd96?w=800&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8OTB8fGNhdHN8ZW58MHx8MHx8fDA%3D',
+  }
+]
+
+export default songs
